Added check for input data

+++ generator/gen.py
@@ -14,7 +14,11 @@ from multiprocessing.dummy import Pool as ThreadPool
 
 models = {
     '1': 'HUAWEI S5320-28P-LI-AC',
-    '2': 'TP-Link T2600G'
+    '2': 'TP-Link T2600G',
+    '3': 'D-Link 1210',
+    '4': 'D-link 3028/1228',
+    '5': 'Eltex',
+    '6': 'Maipu'
 }
 
 items = []  # Для списка с IP
@@ -158,26 +162,33 @@ def convert_switch_data_to_dict(line):
     ranger = list(range(int(list_line[0]), int(list_line[1])+1))
     ports = list(range(1, 25))
     acc_dict = dict(zip(ports, ranger))
-    gateway_list = line[2].split('.')
-    gateway = gateway_list[0] + '.' + \
+    HOSTNAME = line[0]
+    STP = int(float(line[1]))
+    IP = line[2]
+    gateway_list = IP.split('.')
+    GATEWAY = gateway_list[0] + '.' + \
         gateway_list[1] + '.' + gateway_list[2] + '.1'
+    MGMT = int(float(line[4]))
 
-    item = {
-        'HOSTNAME': line[0],
-        'STP': int(line[1]),
-        'IP': line[2],
-        'GATEWAY': gateway,
-        'MGMT': int(line[4]),
-        'access_ports': acc_dict,
-        'B2B_VLAN': int(str(line[4][0])+'00')
-    }
-    full_yaml_file.append(item)
-    items.append(item['IP'])
-    yaml_data_path = script_files_path + 'full_yaml'
-    with open(yaml_data_path, 'w') as final:
-        final.write(yaml.dump(full_yaml_file,
-                              default_flow_style=False, allow_unicode=True))
-    # Возвращает список IP адресов, которые есть в изначальной таблице.
+    if check_ip(IP):
+        item = {
+            'HOSTNAME': HOSTNAME,
+            'STP': STP,
+            'IP': IP,
+            'GATEWAY': GATEWAY,
+            'MGMT': MGMT,
+            'access_ports': acc_dict,
+            'B2B_VLAN': int(str(MGMT)[0]+'00')
+        }
+        full_yaml_file.append(item)
+        items.append(item['IP'])
+        yaml_data_path = script_files_path + 'full_yaml'
+        with open(yaml_data_path, 'w') as final:
+            final.write(yaml.dump(full_yaml_file,
+                                default_flow_style=False, allow_unicode=True))
+    else:
+        print(f'{IP} is wrong. Please check your xls table for correct values of IPs.')
+        pass
     return items
 
 
@@ -191,8 +202,13 @@ def converter(file):
         reader = csv.reader(source)
         next(reader)
         s = list(reader)
+        count = 1
         for switch_data in s:
-            list_of_ip = convert_switch_data_to_dict(switch_data)
+            count += 1
+            try:
+                list_of_ip = convert_switch_data_to_dict(switch_data)
+            except ValueError:
+                print(f'{count} row is empty or wrong.')
     return list_of_ip
 
 
@@ -204,34 +220,30 @@ def config_maker(ip):
     """
 
     yaml_data_path = script_files_path + 'full_yaml'
-    if check_ip(ip):
-        # try:
-        if model == 'HUAWEI S5320-28P-LI-AC':  # huawei
-            env = Environment(loader=FileSystemLoader(
-                'TEMPLATES\\HUAWEI S5320-28P-LI-AC'))
-            if b2b == '2':
-                template = env.get_template('template_uno.txt')
-            else:
-                template = env.get_template('template.txt')
-        elif model == 'TP-Link T2600G':  # tp-link
-            env = Environment(loader=FileSystemLoader(
-                'TEMPLATES\\TP-Link_T2600G'))
+    if model == 'HUAWEI S5320-28P-LI-AC':  # huawei
+        env = Environment(loader=FileSystemLoader(
+            'TEMPLATES\\HUAWEI S5320-28P-LI-AC'))
+        if b2b == '2':
+            template = env.get_template('template_uno.txt')
+        else:
             template = env.get_template('template.txt')
-        with open(yaml_data_path) as full:
-            switches = yaml.safe_load(full)
-        for switch in switches:
-            config_name = switch['IP'] + '.cfg'
-            address = switch['HOSTNAME']
-            address = re.findall(
-                r'([a-zA-Z]+\_+\d+|[a-zA-Z]+\_[a-zA-Z]+\_+\d+)', address)
-            address = address[0]
-            new_path = dir_maker(address, model)
-            config_path = new_path + '\\' + config_name
+    elif model == 'TP-Link T2600G':  # tp-link
+        env = Environment(loader=FileSystemLoader(
+            'TEMPLATES\\TP-Link_T2600G'))
+        template = env.get_template('template.txt')
+    with open(yaml_data_path) as full:
+        switches = yaml.safe_load(full)
+    for switch in switches:
+        config_name = switch['IP'] + '.cfg'
+        address = switch['HOSTNAME']
+        address = re.findall(
+            r'([a-zA-Z]+\_+\d+|[a-zA-Z]+\_[a-zA-Z]+\_+\d+)', address)
+        address = address[0]
+        new_path = dir_maker(address, model)
+        config_path = new_path + '\\' + config_name
 
-            with open(config_path, 'w') as f:
-                f.write(template.render(switch))
-        # except:
-        #   print('IP-address error. Please check your xls table for correct values of IPs.')
+        with open(config_path, 'w') as f:
+            f.write(template.render(switch))
 
 
 def model_choicer(model):
@@ -276,33 +288,33 @@ def script_files():
         pass
     return path_for_res
 
-
-def input_checker():
-    """
-    Обрабатывает ввод команд на предмет корректности. Проверяет существование файла, проверяет корректный выбор коммутатора. 
-    
-    """
-    pass
-
+def filename_exist_checker(filename):
+    os.path.exists(filename)
 
 if __name__ == '__main__':
-    table_of_switches = input('Enter filename: ')
-    b2b = str(input('For common enter 1\nFor b2b enter 2\nEnter:'))
+    
+    while True:
+        table_of_switches = 'serova_tp.xls'
+        #table_of_switches = input('Enter filename: ')
+        if os.path.exists(table_of_switches):
+            b2b = str(input('For common enter 1\nFor b2b enter 2\nEnter:'))
+            #table_of_switches = 'serova_tp.xls'
+            for k, v in models.items():
+                print(f'To make {v} configuration enter {k}')
+            switch_model = input("Enter: ")
+            script_files_path = script_files() + '\\'
 
-    #table_of_switches = 'nov.xls'
-    for k, v in models.items():
-        print(f'To make {v} configuration enter {k}')
-    switch_model = input("Enter: ")
-    script_files_path = script_files() + '\\'
+            model = model_choicer(str(switch_model))
+            csv_file_path = csv_from_excel(table_of_switches)
+            transliterator(csv_file_path)
+            ip_list = converter(csv_file_path)
 
-    model = model_choicer(str(switch_model))
-    csv_file_path = csv_from_excel(table_of_switches)
-    transliterator(csv_file_path)
-    ip_list = converter(csv_file_path)
-
-    pool = ThreadPool(2)
-    pool.map(config_maker, ip_list)
-    pool.close()
-    pool.join()
-    shutil.rmtree(script_files_path)
-    input("Done. Press 'Enter' for exit.")
+            pool = ThreadPool(2)
+            pool.map(config_maker, ip_list)
+            pool.close()
+            pool.join()
+            shutil.rmtree(script_files_path)
+            input("Done. Press 'Enter' for exit.")
+            break
+        else:
+            print(f'"{table_of_switches}" is not found!\n')
